Larger test cleanup
